meson: Make features fully configurable
This commit enables a (mostly automatic) way to configure ffmpeg with a
reduced feature set. This maps closely what's provided with upstream's
configure script, but without the ability of mapping components to a
given category, yet.

To do this, a new flag `--meson-options` has been added to
find_things_extern.py, that will output the found components and their
corresponding option() entries.

These have been manually added for the existing 4.4.2 release, and will
need to be checked for every update.